fix some of the warnings from Intel compiler and fix bug getting entities by dense tag value
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4352 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ src/moab/CN.hpp
@@ -579,7 +579,7 @@ inline void CN::setPermutation(const EntityType t, const int dim, short int *pve
     that_vec[pvec[i]] = i;
   }
 
-  this_vec[MAX_SUB_ENTITIES] = that_vec[MAX_SUB_ENTITIES] = num_entries;
+  this_vec[MAX_SUB_ENTITIES] = that_vec[MAX_SUB_ENTITIES] = (short)num_entries;
 }
 
 //! Reset permutation or reverse permutation vector
